[REFACTOR] Fix issue on order of singleton services and Implemented SRP Design Pattern in nutrition page
