+++ test/features/food_scan_ai/presentation/food_scan_page_test.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:pockeat/features/food_scan_ai/presentation/food_scan_page.dart';
-import 'package:pockeat/features/food_scan_ai/presentation/nutrition_page.dart';
+import 'package:pockeat/features/food_scan_ai/presentation/pages/food_scan_page.dart';
+import 'package:pockeat/features/food_scan_ai/presentation/pages/nutrition_page.dart';
 import 'package:camera/camera.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:get_it/get_it.dart';
